refactor component imports and remove unused sidebar components
