--- resources/js/components/ui/sidebar/SidebarGroupAction.vue
@@ -1,31 +1,0 @@
-<script setup lang="ts">
-import { cn } from '@/lib/utils';
-import type { PrimitiveProps } from 'reka-ui';
-import { Primitive } from 'reka-ui';
-import type { HTMLAttributes } from 'vue';
-
-const props = defineProps<
-    PrimitiveProps & {
-        class?: HTMLAttributes['class'];
-    }
->();
-</script>
-
-<template>
-    <Primitive
-        data-slot="sidebar-group-action"
-        data-sidebar="group-action"
-        :as="as"
-        :as-child="asChild"
-        :class="
-            cn(
-                'absolute top-3.5 right-3 flex aspect-square w-5 items-center justify-center rounded-md p-0 text-sidebar-foreground ring-sidebar-ring outline-hidden transition-transform hover:bg-sidebar-accent hover:text-sidebar-accent-foreground focus-visible:ring-2 [&>svg]:size-4 [&>svg]:shrink-0',
-                'after:absolute after:-inset-2 md:after:hidden',
-                'group-data-[collapsible=icon]:hidden',
-                props.class,
-            )
-        "
-    >
-        <slot />
-    </Primitive>
-</template>
