--- resources/js/components/ui/sidebar/SidebarGroupLabel.vue
@@ -1,30 +1,0 @@
-<script setup lang="ts">
-import { cn } from '@/lib/utils';
-import type { PrimitiveProps } from 'reka-ui';
-import { Primitive } from 'reka-ui';
-import type { HTMLAttributes } from 'vue';
-
-const props = defineProps<
-    PrimitiveProps & {
-        class?: HTMLAttributes['class'];
-    }
->();
-</script>
-
-<template>
-    <Primitive
-        data-slot="sidebar-group-label"
-        data-sidebar="group-label"
-        :as="as"
-        :as-child="asChild"
-        :class="
-            cn(
-                'flex h-8 shrink-0 items-center rounded-md px-2 text-xs font-medium text-sidebar-foreground/70 ring-sidebar-ring outline-hidden transition-[margin,opacity] duration-200 ease-linear focus-visible:ring-2 [&>svg]:size-4 [&>svg]:shrink-0',
-                'group-data-[collapsible=icon]:-mt-8 group-data-[collapsible=icon]:opacity-0',
-                props.class,
-            )
-        "
-    >
-        <slot />
-    </Primitive>
-</template>
