--- resources/js/components/ui/sidebar/SidebarMenuSub.vue
@@ -1,24 +1,0 @@
-<script setup lang="ts">
-import { cn } from '@/lib/utils';
-import type { HTMLAttributes } from 'vue';
-
-const props = defineProps<{
-    class?: HTMLAttributes['class'];
-}>();
-</script>
-
-<template>
-    <ul
-        data-slot="sidebar-menu-sub"
-        data-sidebar="menu-badge"
-        :class="
-            cn(
-                'mx-3.5 flex min-w-0 translate-x-px flex-col gap-1 border-l border-sidebar-border px-2.5 py-0.5',
-                'group-data-[collapsible=icon]:hidden',
-                props.class,
-            )
-        "
-    >
-        <slot />
-    </ul>
-</template>
